pyproject.toml
[project]
name = "proveai-supply-chain-simulation"
version = "0.1.0"
description = "Two-stage supply-chain negotiation simulator with Anthropic agents, LangGraph orchestration, Langfuse tracing, and a Streamlit dashboard."
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "anthropic==0.94.0",
    "langgraph==1.1.6",
    "langfuse==4.2.0",
    "streamlit==1.56.0",
    "python-dotenv==1.2.2",
    "plotly==6.7.0",
]

[tool.uv]
package = false
